% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardupilot_log_functions.R
\name{class.define.list}
\alias{class.define.list}
\title{Apply define class to dataframes in a list}
\usage{
class.define.list(x)
}
\arguments{
\item{x}{A list of dataframes}
}
\value{
The list of dataframes with defined integer/numeric classes
}
\description{
Apply define class to dataframes in a list
}
