% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardupilot_log_functions.R
\name{class.define}
\alias{class.define}
\title{Define class}
\usage{
class.define(x)
}
\arguments{
\item{x}{A vector containing a message parameter}
}
\value{
The vector with the class defined as either integer or numeric
}
\description{
Determines if vectors is numeric or integer to minimize memory
}
\examples{
x <- c(1,2,4,5)
str(x)
x <- class.define(x)
str(x)
}
