% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotagging_functions.R
\name{geodesic.distance}
\alias{geodesic.distance}
\title{Calculate the great-circle distance}
\usage{
geodesic.distance(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{Longititude of point 1}

\item{lat1}{Latitude ot point 1}

\item{lon2}{Longititude of point 2}

\item{lat2}{Latitude ot point 2}
}
\value{
Distance in meters
}
\description{
\code{geodesic.distance} calculates the great-circle distance between two points
}
\examples{
# Distance from Toronto to Ottawa in m
dist_otttor <- geodesic.distance(-75.695000, 45.424721, -79.347015, 43.651070)
# in km
dist_ottor / 1000
}
