% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardupilot_log_functions.R
\name{gps.time.to.utc}
\alias{gps.time.to.utc}
\title{Convert GPS time to UTC time}
\usage{
gps.time.to.utc(GPSweek, GPSms, hrsDiff = 0, leapsec = 17)
}
\arguments{
\item{GPSweek}{GPS week number - ardupilot log GPS message parameter (GWk)}

\item{GPSms}{GPS time in milliseconds - ardupilot log GPS message parameter (GMS) -}

\item{hrsDiff}{May set time difference in hours to UTC time of your location}

\item{leapsec}{Time correction between GPS and UTC time in seconds}
}
\value{
Time in UTC time unless included hours difference argument
}
\description{
\code{GPStimeToUTC} converts GPS time in weeks and milliseconds to UTC time
}
\examples{
gps_week <- 1923
gps_seconds <- 381765
gps_milliseconds <- gps_seconds * 1000
utc_time <- GPStimeToUTC(gps_week, gps_milliseconds)
utc_time
}
