% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardupilot_log_functions.R
\name{read.apm.log}
\alias{read.apm.log}
\title{Import ardupilot dataflash log}
\usage{
read.apm.log(file)
}
\arguments{
\item{file}{an ardupilot dataflash log (.log)}
}
\value{
The ardupilot messages are organized into a list of dataframes
  containing columns for the message parameters
}
\description{
Import ardupilot dataflash log
}
