% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotagging_functions.R
\name{read.kap.log}
\alias{read.kap.log}
\title{Reads KAP UAV Exposure Control Script Logged Data}
\usage{
read.kap.log(file)
}
\arguments{
\item{file}{a KAP.log file}
}
\value{
A dataframe containing the date, time, image name, shutter speed (Tv),
 aperature setting (Av), ISO sensitivity (Sv), and time convered to seconds.
}
\description{
\code{read.kap.log} imports the logged camera data from the KAP UAV Exposure
 Control Script.
}
