% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotagging_functions.R
\name{read.rtklib.pos}
\alias{read.rtklib.pos}
\title{Imports a processed RTKLIB file}
\usage{
read.rtklib.pos(file, skip = 26)
}
\arguments{
\item{file}{an RTKLIB.pos file}

\item{skip}{how many lines to skip header info - should probably refine this}
}
\value{
A dataframe with GNSS positional information from the RTKLIB pos file
}
\description{
\code{read.rtklib.pos} imports a processed RTKLIB file http://www.rtklib.com
}
