% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotagging_functions.R
\name{select.photos}
\alias{select.photos}
\title{Select Sequence of Photos from Kap.log}
\usage{
select.photos(kapLog, start, end)
}
\arguments{
\item{kapLog}{the dataframe of the importated KAP.log file}

\item{start}{the name (image.name) of the first file in a sequence}

\item{end}{the name (image.name) of the last file in a sequence}
}
\value{
A dataframe containing the KAP log info for a sequence of photos
}
\description{
\code{select.photos} selects a sequence of photos from the imported KAP.log
}
