% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotagging_functions.R
\name{time.to.seconds}
\alias{time.to.seconds}
\title{Convert time in 24h to seconds}
\usage{
\method{time}{to.seconds}(x)
}
\arguments{
\item{x}{time string (00:00:00.000)}
}
\value{
Time in seconds
}
\description{
\code{time.to.seconds} converts time from a 24h to seconds
}
\examples{
time.to.seconds("10:44:13.133")
}
